+++ packages/@webex/internal-plugin-metrics/src/call-diagnostic/call-diagnostic-metrics.util.ts
@@ -246,11 +246,13 @@ export const prepareDiagnosticMetricItem = (webex: any, item: any) => {
       break;
     case 'client.login.end':
       joinTimes.otherAppApiReqResp = cdl.getOtherAppApiReqResp();
+      joinTimes.exchangeCITokenJMT = cdl.getExchangeCITokenJMT();
       break;
     case 'client.interstitial-window.launched':
       joinTimes.meetingInfoReqResp = cdl.getMeetingInfoReqResp();
       joinTimes.clickToInterstitial = cdl.getClickToInterstitial();
       joinTimes.refreshCaptchaServiceReqResp = cdl.getRefreshCaptchaReqResp();
+      joinTimes.downloadIntelligenceModelsReqResp = cdl.getDownloadIntelligenceModelsReqResp();
       break;
 
     case 'client.call.initiated':
@@ -258,6 +260,7 @@ export const prepareDiagnosticMetricItem = (webex: any, item: any) => {
       joinTimes.showInterstitialTime = cdl.getShowInterstitialTime();
       joinTimes.registerWDMDeviceJMT = cdl.getRegisterWDMDeviceJMT();
       joinTimes.getU2CTime = cdl.getU2CTime();
+      joinTimes.getReachabilityClustersReqResp = cdl.getReachabilityClustersReqResp();
       break;
 
     case 'client.locus.join.response':