+++ packages/@webex/internal-plugin-metrics/test/unit/spec/call-diagnostic/call-diagnostic-metrics-batcher.ts
@@ -56,7 +56,7 @@ describe('plugin-metrics', () => {
     });
 
     describe('#request()', () => {
-      describe('when the request completes successfully', async () => {
+      describe('when the request completes successfully', () => {
         it('clears the queue', async () => {
           const promise = webex.internal.newMetrics.callDiagnosticMetrics.submitToCallDiagnostics(
             //@ts-ignore
@@ -106,6 +106,9 @@ describe('plugin-metrics', () => {
           webex.internal.newMetrics.callDiagnosticLatencies.getRefreshCaptchaReqResp = sinon
             .stub()
             .returns(10);
+          webex.internal.newMetrics.callDiagnosticLatencies.getDownloadIntelligenceModelsReqResp =
+            sinon.stub().returns(42);
+
           const promise = webex.internal.newMetrics.callDiagnosticMetrics.submitToCallDiagnostics(
             //@ts-ignore
             {event: {name: 'client.interstitial-window.launched'}}
@@ -123,6 +126,7 @@ describe('plugin-metrics', () => {
               clickToInterstitial: 10,
               meetingInfoReqResp: 10,
               refreshCaptchaServiceReqResp: 10,
+              downloadIntelligenceModelsReqResp: 42,
             },
           });
           assert.lengthOf(
@@ -138,6 +142,9 @@ describe('plugin-metrics', () => {
           webex.internal.newMetrics.callDiagnosticLatencies.getU2CTime = sinon
             .stub()
             .returns(20);
+          webex.internal.newMetrics.callDiagnosticLatencies.getReachabilityClustersReqResp = sinon
+            .stub()
+            .returns(10);
           const promise = webex.internal.newMetrics.callDiagnosticMetrics.submitToCallDiagnostics(
             //@ts-ignore
             {event: {name: 'client.call.initiated'}}
@@ -154,7 +161,8 @@ describe('plugin-metrics', () => {
               meetingInfoReqResp: 10,
               registerWDMDeviceJMT: 10,
               showInterstitialTime: 10,
-              getU2CTime: 20
+              getU2CTime: 20,
+              getReachabilityClustersReqResp: 10
             },
           });
           assert.lengthOf(
@@ -175,7 +183,7 @@ describe('plugin-metrics', () => {
           webex.internal.newMetrics.callDiagnosticLatencies.getCallInitJoinReq = sinon
             .stub()
             .returns(10);
-            webex.internal.newMetrics.callDiagnosticLatencies.getDownloadTimeJMT = sinon
+          webex.internal.newMetrics.callDiagnosticLatencies.getDownloadTimeJMT = sinon
             .stub()
             .returns(100);
           const promise = webex.internal.newMetrics.callDiagnosticMetrics.submitToCallDiagnostics(