+++ packages/@webex/internal-plugin-metrics/test/unit/spec/call-diagnostic/call-diagnostic-metrics.util.ts
@@ -40,7 +40,6 @@ describe('internal-plugin-metrics', () => {
         },
       };
       clearEmptyKeysRecursively(obj);
-      console.log(obj);
       assert.deepEqual(obj, {nested: {arr: ['test']}});
     });
 
@@ -261,7 +260,7 @@ describe('internal-plugin-metrics', () => {
     });
   });
 
-  describe('prepareDiagnosticMetricItem', async () => {
+  describe('prepareDiagnosticMetricItem', () => {
     let webex: any;
 
     const check = (eventName: string, expectedEvent: any) => {
@@ -283,7 +282,7 @@ describe('internal-plugin-metrics', () => {
       });
     };
 
-    before(async () => {
+    beforeEach(async () => {
       webex = {internal: {newMetrics: {}}};
       webex.internal.newMetrics.callDiagnosticLatencies = new CallDiagnosticLatencies(
         {},
@@ -305,6 +304,7 @@ describe('internal-plugin-metrics', () => {
       ['client.login.end', {
         joinTimes: {
           otherAppApiReqResp: undefined,
+          exchangeCITokenJMT: undefined,
         }
       }],
       ['client.webexapp.launched', {
@@ -319,6 +319,7 @@ describe('internal-plugin-metrics', () => {
             clickToInterstitial: undefined,
             meetingInfoReqResp: undefined,
             refreshCaptchaServiceReqResp: undefined,
+            downloadIntelligenceModelsReqResp: undefined,
           },
         },
       ],
@@ -329,7 +330,8 @@ describe('internal-plugin-metrics', () => {
             showInterstitialTime: undefined,
             meetingInfoReqResp: undefined,
             registerWDMDeviceJMT: undefined,
-            getU2CTime: undefined
+            getU2CTime: undefined,
+            getReachabilityClustersReqResp: undefined,
           },
         },
       ],
@@ -345,7 +347,7 @@ describe('internal-plugin-metrics', () => {
             interstitialToJoinOK: undefined,
             totalJmt: undefined,
             clientJmt: undefined,
-            downloadTime: undefined
+            downloadTime: undefined,
           },
         },
       ],
@@ -415,7 +417,7 @@ describe('internal-plugin-metrics', () => {
     });
   });
 
-  describe('setMetricTimings', async () => {
+  describe('setMetricTimings', () => {
     let webex: any;
 
     const check = (options: any, expectedOptions: any) => {